backend/requirements.txt
flask
flask-cors
pymongo[srv]
pyjwt
werkzeug
python-dotenv
pandas
openpyxl
pytz
Pillow
pytesseract
pdfplumber